requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
boto3==1.28.28
    # via sqlcritic (pyproject.toml)
botocore==1.31.28
    # via
    #   boto3
    #   s3transfer
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via pyjwt
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   pygithub
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via opentelemetry-api
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
opentelemetry-api==1.19.0
    # via
    #   opentelemetry-sdk
    #   sqlcritic (pyproject.toml)
opentelemetry-sdk==1.19.0
    # via sqlcritic (pyproject.toml)
opentelemetry-semantic-conventions==0.40b0
    # via opentelemetry-sdk
pycparser==2.21
    # via cffi
pygithub==1.59.1
    # via sqlcritic (pyproject.toml)
pyjwt[crypto]==2.8.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
python-dateutil==2.8.2
    # via botocore
requests==2.31.0
    # via pygithub
s3transfer==0.6.2
    # via boto3
six==1.16.0
    # via python-dateutil
typing-extensions==4.7.1
    # via opentelemetry-sdk
urllib3==1.26.16
    # via
    #   botocore
    #   requests
wrapt==1.15.0
    # via deprecated
zipp==3.16.2
    # via importlib-metadata
